fix(review): move shared test helpers _make_tool_use_line/_sr/_success_session_json to conftest
Cross-test-module imports from test_headless_core.py (private underscore helpers)
were fragile under xdist — renaming or signature changes break sibling files at
import time. Centralizing them in tests/execution/conftest.py removes the coupling.
